Moved MainFormDataTest and MainFormWrapperTest to proper locations to match corresponding Main class.  Refactored MainFormData and Test to use constants for property change keys
